+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # Automated Tourism
+## v0.04 05/04/2019 9:52:28 PM
+
+#### Added
+- local function ATsetupVariables(rocket, init)
+- local Old_SupplyRocket_OnDemolish = SupplyRocket.OnDemolish
+- Finished ip status section
+
+#### Todo
+
+- on screen notifications
+
+--------------------------------------------------------
 ## v0.03 05/04/2019 5:03:27 PM
 #### Changed
 - disable all buttons except priority